Move views into their own apps
