+++ .idea/misc.xml
@@ -20,7 +20,7 @@
   <component name="ProjectModuleManager">
     <modules />
   </component>
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 2.7.8 (/Library/Frameworks/Python.framework/Versions/2.7/bin/python2.7)" project-jdk-type="Python SDK" />
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.3.5 (/Library/Frameworks/Python.framework/Versions/3.3/bin/python3.3)" project-jdk-type="Python SDK" />
   <component name="RunManager">
     <list size="0" />
   </component>
