+++ .idea/modules.xml
@@ -2,7 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/myproject.iml" filepath="$PROJECT_DIR$/.idea/myproject.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/my-mez.iml" filepath="$PROJECT_DIR$/.idea/my-mez.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/my-mez.iml" filepath="$PROJECT_DIR$/.idea/my-mez.iml" />
     </modules>
   </component>
 </project>
