+++ mysite/Team_Images/views.py
